Eliminate stage1_find_marks_flatten.h
